backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg[binary]==3.1.18
django-cors-headers==4.3.1
razorpay==1.4.2
Pillow==10.2.0
reportlab==4.0.9
python-decouple==3.8
PyJWT==2.8.0